requirements.txt
requests
pandas
python-dotenv
Pillow
aiohttp